pyproject.toml
[project]
name = "grove-sync"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "supabase",
    "pyyaml",
    "requests",
    "boto3>=1.42.74",
    "python-dateutil>=2.9.0.post0",
    "yt-dlp>=2025.3.31",
    "curl-cffi>=0.14.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[project.scripts]
sync-plan = "src.commands.plan:main"
sync-process = "src.commands.process:main"
sync-calibrate = "src.commands.calibrate:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.format]
indent-style = "space"
quote-style = "double"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade
]
ignore = []
preview = true

[tool.mypy]
python_version = "3.12"
namespace_packages = true
explicit_package_bases = true
packages = ["src"]

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests checking full behavior",
    "safety: Tests that verify safety constraints and prevent dangerous operations",
    "edge_cases: Tests for edge cases and boundary conditions",
]
